requirements.txt
fastapi>=0.95.0
uvicorn[standard]
pytest
httpx
motor
python-dotenv
PyJWT
email-validator
